+++ aravis/src/manual/mod.rs
@@ -18,8 +18,11 @@ pub(crate) mod traits {
 	pub use super::register::traits::GcRegisterExtManual;
 }
 
+#[cfg(feature = "image")]
 pub type BoxImage<P> = image::ImageBuffer<P, Box<[u8]>>;
+#[cfg(feature = "image")]
 pub type RcImage<P> = image::ImageBuffer<P, std::rc::Rc<[u8]>>;
+#[cfg(feature = "image")]
 pub type ArcImage<P> = image::ImageBuffer<P, std::sync::Arc<[u8]>>;
 
 static INITIALIZED: AtomicBool = AtomicBool::new(false);